teste

+++ meupet/src/main/java/com/br/unisales/service/PetService.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import com.br.unisales.configuration.ConfigurationManager;
 import com.br.unisales.table.Pet;
+import com.br.unisales.table.Proprietario;
 
 import jakarta.persistence.TypedQuery;
 
@@ -22,18 +23,28 @@ public class PetService {
             return query.getResultList();
         } catch (Exception e) {          
             System.err.println("Erro: " + e.getMessage());
-            return new ArrayList<Pet>();
+            return new ArrayList<>();
         }
     }
 
     public Pet buscarPorId(Integer id) {
-        TypedQuery<Pet> query = this.config.getEntityManager().createQuery("FROM Pet WHERE id = :id", Pet.class);
-        query.setParameter("id", id);
-        return query.getSingleResult();
+        try {
+            TypedQuery<Pet> query = this.config.getEntityManager().createQuery("FROM Pet WHERE id = :id", Pet.class);
+            query.setParameter("id", id);
+            return query.getSingleResult();
+        } catch (Exception e) {
+            System.err.println("Erro: " + e.getMessage());
+            return null;
+        }
     }
 
     public Pet salvar(Integer id, String nome, Double peso, String raca) {
-        Pet pet = new Pet(id, nome,  peso, raca);
+        Pet pet = Pet.builder()
+                     .id(id)
+                     .nome(nome)
+                     .peso(peso)
+                     .raca(raca)
+                     .build();
         this.config.getEntityManager().getTransaction().begin();
         if(id == null) {
             this.config.getEntityManager().persist(pet);
@@ -48,10 +59,12 @@ public class PetService {
     public String excluir(Integer id) {
         if(id != null) {
             Pet pet = this.buscarPorId(id);
-            this.config.getEntityManager().getTransaction().begin();
-            this.config.getEntityManager().remove(pet);
-            this.config.getEntityManager().getTransaction().commit();
-            return "ok";
+            if (pet != null) {
+                this.config.getEntityManager().getTransaction().begin();
+                this.config.getEntityManager().remove(pet);
+                this.config.getEntityManager().getTransaction().commit();
+                return "ok";
+            }
         }
         
         return "erro";
